refactor: remove dead exports, reduce indirection, clean up legacy code
- Remove dead exports from package-versions.ts: resolveVersionManifest,
  VersionDetail interface, CreateVersionError/CreateVersionResult types
- Un-export internal webhook symbols: webhookEventSchema, webhookEventsSchema,
  WebhookEventType, validateWebhookUrl, buildSignedHeaders (keep validateEvents
  exported for test coverage)
- Move isPromptEmpty + findMissingDependencies from @appstrate/shared-types
  to @appstrate/core/validation where manifest logic belongs
- Extract extractManifestSchemas() helper in manifest-utils.ts — replaces
  duplicated input/config/output schema extraction across env-builder and
  agent-readiness
- Remove legacy trigger cleanup from notify.ts (15 lines of DROP TRIGGER/
  FUNCTION for pre-migration triggers — no production data exists)
