+++ packages/core/src/validation.ts
@@ -477,3 +477,24 @@ export function validateToolSource(source: string): ToolSourceValidationResult {
 
   return { valid: errors.length === 0, errors, warnings };
 }
+
+// ─────────────────────────────────────────────
+// Agent readiness utilities
+// ─────────────────────────────────────────────
+
+/** Check if a prompt is empty or whitespace-only. */
+export function isPromptEmpty(prompt: string): boolean {
+  return prompt.trim().length === 0;
+}
+
+/**
+ * Find IDs declared in `required` but missing from `installed`.
+ * Works for both skills and tools.
+ */
+export function findMissingDependencies(
+  required: Record<string, string>,
+  installedIds: string[],
+): string[] {
+  const installed = new Set(installedIds);
+  return Object.keys(required).filter((id) => !installed.has(id));
+}
